Now progressbar works in threading mode (thus also for serial jobs). Moved the code to lib.log
